+++ cmd/kava/cmd/root.go
@@ -22,6 +22,7 @@ import (
 	"github.com/kava-labs/kava/app"
 	"github.com/kava-labs/kava/app/params"
 	kavaclient "github.com/kava-labs/kava/client"
+	"github.com/kava-labs/kava/cmd/kava/cmd/rocksdb"
 	"github.com/kava-labs/kava/cmd/kava/opendb"
 	"github.com/kava-labs/kava/migrate"
 )
@@ -122,5 +123,6 @@ func addSubCmds(rootCmd *cobra.Command, encodingConfig params.EncodingConfig, de
 		kavaclient.KeyCommands(app.DefaultNodeHome),
 		newIavlViewerCmd(opts),
 		newShardCmd(opts),
+		rocksdb.RocksDBCmd,
 	)
 }