+++ CHANGELOG.md
@@ -36,8 +36,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Features
+
+- (cli) [#1785] Add `shard` CLI command to support creating partitions of data for standalone nodes
 - (metrics) [#1668] Adds non-state breaking x/metrics module for custom telemetry.
 - (metrics) [#1669] Add performance timing metrics to all Begin/EndBlockers
+
+### Bug Fixes
 - (rocksdb) [#1767] Fix resolution of rocksdb database path
 
 ## [v0.21.1](https://github.com/Kava-Labs/kava/releases/tag/v0.21.1)
@@ -211,6 +216,7 @@ Bump tendermint version to 0.32.10 to address [cosmos security advisory Lavender
 [\#257](https://github.com/Kava-Labs/kava/pulls/257) Include scripts to run large-scale simulations remotely using aws-batch
 
 
+[#1785]: https://github.com/Kava-Labs/kava/pull/1785
 [#1767]: https://github.com/Kava-Labs/kava/pull/1767
 [#1669]: https://github.com/Kava-Labs/kava/pull/1669
 [#1668]: https://github.com/Kava-Labs/kava/pull/1668