+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (metrics) [#1668] Adds non-state breaking x/metrics module for custom telemetry.
 - (metrics) [#1669] Add performance timing metrics to all Begin/EndBlockers
 - (cli) [#1785] Add `shard` CLI command to support creating partitions of data for standalone nodes
+- (cli) [#1785] Add `shard` CLI command to support creating partitions of data for standalone nodes
 
 ### Bug Fixes
 - (rocksdb) [#1767] Fix resolution of rocksdb database path
@@ -172,6 +173,7 @@ Bump tendermint version to 0.32.10 to address [cosmos security advisory Lavender
 
 [\#257](https://github.com/Kava-Labs/kava/pulls/257) Include scripts to run large-scale simulations remotely using aws-batch
 
+[#1804]: https://github.com/Kava-Labs/kava/pull/1804
 [#1785]: https://github.com/Kava-Labs/kava/pull/1785
 [#1767]: https://github.com/Kava-Labs/kava/pull/1767
 [#1669]: https://github.com/Kava-Labs/kava/pull/1669